backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
pgvector==0.2.4
alembic==1.13.1

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & Embeddings
google-generativeai==0.3.2
openai==1.10.0

# Document Processing
python-docx==1.1.0
openpyxl==3.1.2
reportlab==4.0.9
PyPDF2==3.0.1

# GCP
google-cloud-storage==2.14.0

# Utils
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
